refactor: create youtubeData utils
