+++ src/components/Link/LinkForm.tsx
@@ -13,10 +13,10 @@ import useSWR from 'swr';
 
 import { getLinkDetails, LINK_URL } from '@/apis/Media';
 import useCategory from '@/hooks/useCategory';
-import { useYoutubeVideo } from '@/hooks/UseYoutubeVideo';
 import { CategoryResponseDTO } from '@/types/category.interface';
 import { FormData } from '@/types/media.interface';
 import { getLinkUrlInfo } from '@/utils/validations/linkUtils';
+import { useYoutubeVideo } from '@/utils/youtubeData';
 
 interface LinkFormProps {
   onSubmit: (data: FormData) => void;
