+++ src/utils/youtubeData.ts
@@ -1,12 +1,36 @@
 import { useCallback, useState } from 'react';
 
 import { fetchYoutubeVideo } from '@/apis/Youtube';
-import {
-  getYoutubeVideoFromMetadata,
-  YoutubeVideo,
-} from '@/hooks/useGetYoutubeVideoFromMetadata';
 
-export function extractVideoIdFromUrl(url: string) {
+interface YoutubeVideo {
+  title: string;
+  thumbnailUrl: string;
+  description: string;
+}
+
+export interface YoutubeVideoMetadata {
+  snippet: {
+    title: string;
+    thumbnails?: {
+      default?: {
+        url: string;
+      };
+    };
+    description: string;
+  };
+}
+
+function getYoutubeVideoFromMetadata(
+  youtubeVideoMetadata: YoutubeVideoMetadata
+): YoutubeVideo {
+  return {
+    title: youtubeVideoMetadata.snippet.title,
+    thumbnailUrl: youtubeVideoMetadata.snippet.thumbnails?.default?.url || '',
+    description: youtubeVideoMetadata.snippet.description.slice(0, 500),
+  };
+}
+
+function extractVideoIdFromUrl(url: string) {
   const match = url.match(/(?:[?&]v=|\/embed\/|\/v\/|youtu\.be\/)([^&\n?#]+)/);
   return match ? match[1] : '';
 }
